test: organize fixtures within a subdirectory
