+++ tests/conftest.py
@@ -5,8 +5,6 @@ from brownie.project.main import get_loaded_projects
 from pathlib import Path
 
 from brownie_hooks import DECIMALS as hook_decimals
-from scripts.utils import right_pad, pack_values
-
 
 # functions in wrapped methods are renamed to simplify common tests
 
@@ -25,7 +23,13 @@ WRAPPED_COIN_METHODS = {
     },
 }
 
-ZERO_ADDRESS = "0x0000000000000000000000000000000000000000"
+pytest_plugins = [
+    "fixtures.accounts",
+    "fixtures.deployments",
+    "fixtures.functions",
+    "fixtures.pooldata",
+    "fixtures.setup",
+]
 
 _pooldata = {}
 
@@ -84,7 +88,7 @@ def pytest_generate_tests(metafunc):
     if "pool_data" in metafunc.fixturenames:
         # parametrize `pool_data`
         test_path = Path(metafunc.definition.fspath).relative_to(project._path)
-        if test_path.parts[1] == "common":
+        if test_path.parts[:3] == ("tests", "pools", "common"):
             if metafunc.config.getoption("pool"):
                 params = metafunc.config.getoption("pool").split(',')
             else:
@@ -145,7 +149,7 @@ def isolation_setup(fn_isolation):
 
 # main parametrized fixture, used to pass data about each pool into the other fixtures
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 def pool_data(request):
     project = get_loaded_projects()[0]
 
@@ -153,197 +157,11 @@ def pool_data(request):
         pool_name = request.param
     else:
         test_path = Path(request.fspath).relative_to(project._path)
-        pool_name = test_path.parts[1]
+        # ("tests", "pools", pool_name, ...)
+        pool_name = test_path.parts[2]
     yield _pooldata[pool_name]
-
-
-# pool-dependent data fixtures
-
-@pytest.fixture(scope="module")
-def underlying_decimals(pool_data):
-    # number of decimal places for each underlying coin in the active pool
-    yield [i['decimals'] for i in pool_data['coins']]
-
-
-@pytest.fixture(scope="module")
-def wrapped_decimals(pool_data):
-    # number of decimal places for each wrapped coin in the active pool
-    yield [i.get('wrapped_decimals', i['decimals']) for i in pool_data['coins']]
-
-
-@pytest.fixture(scope="module")
-def initial_amounts(wrapped_decimals):
-    # 1e6 of each coin - used to make an even initial deposit in many test setups
-    yield [10**(i+6) for i in wrapped_decimals]
-
-
-@pytest.fixture(scope="module")
-def n_coins(pool_data):
-    yield len(pool_data['coins'])
-
-
-# named accounts
-
-@pytest.fixture(scope="session")
-def alice(accounts):
-    yield accounts[0]
-
-
-@pytest.fixture(scope="session")
-def bob(accounts):
-    yield accounts[1]
-
-
-@pytest.fixture(scope="session")
-def charlie(accounts):
-    yield accounts[2]
 
 
 @pytest.fixture(scope="session")
 def project():
     yield get_loaded_projects()[0]
-
-
-# contract deployments
-
-@pytest.fixture(scope="module")
-def wrapped_coins(project, alice, pool_data, underlying_coins):
-    if not pool_data.get("wrapped_contract"):
-        yield underlying_coins
-    else:
-        fn_names = WRAPPED_COIN_METHODS[pool_data['wrapped_contract']]
-        deployer = getattr(project, pool_data['wrapped_contract'])
-        coins = []
-        for i, coin_data in enumerate(pool_data['coins']):
-            underlying = underlying_coins[i]
-            if not coin_data['wrapped']:
-                coins.append(underlying)
-            else:
-                decimals = coin_data['wrapped_decimals']
-                contract = deployer.deploy(
-                    f"Coin {i}", f"C{i}", decimals, 0, underlying, 10**18, {'from': alice}
-                )
-                for target, attr in fn_names.items():
-                    setattr(contract, target, getattr(contract, attr))
-                coins.append(contract)
-
-        yield coins
-
-
-@pytest.fixture(scope="module")
-def underlying_coins(ERC20Mock, ERC20MockNoReturn, alice, pool_data):
-    coins = []
-    for i, coin_data in enumerate(pool_data['coins']):
-        decimals = coin_data['decimals']
-        deployer = ERC20MockNoReturn if coin_data['tethered'] else ERC20Mock
-        contract = deployer.deploy(f"Underlying Coin {i}", f"UC{i}", decimals, {'from': alice})
-        coins.append(contract)
-
-    yield coins
-
-
-@pytest.fixture(scope="module")
-def pool_token(project, alice, pool_data):
-    name = pool_data['name']
-    deployer = getattr(project, pool_data['lp_contract'])
-    yield deployer.deploy(f"Curve {name} LP Token", f"{name}CRV", 18, 0, {'from': alice})
-
-
-@pytest.fixture(scope="module")
-@pytest.mark.parametrize()
-def swap(project, alice, underlying_coins, wrapped_coins, pool_token, pool_data):
-    deployer = getattr(project, pool_data['swap_contract'])
-
-    abi = next(i['inputs'] for i in deployer.abi if i['type'] == "constructor")
-    args = {
-        '_coins': wrapped_coins,
-        '_underlying_coins': underlying_coins,
-        '_pool_token': pool_token,
-        '_A': 360 * 2,
-        '_fee': 0,
-        '_admin_fee': 0,
-        '_offpeg_fee_multiplier': 0,
-        '_owner': alice,
-    }
-    deployment_args = [args[i['name']] for i in abi] + [({'from': alice})]
-
-    contract = deployer.deploy(*deployment_args)
-    pool_token.set_minter(contract, {'from': alice})
-
-    yield contract
-
-
-@pytest.fixture(scope="module")
-def registry(
-    Registry,
-    alice,
-    gauge_controller,
-    swap,
-    pool_token,
-    n_coins,
-    wrapped_coins,
-    wrapped_decimals,
-    underlying_decimals,
-    pool_data,
-):
-    registry = Registry.deploy(gauge_controller, {'from': alice})
-
-    rate_sig = "0x00"
-    if next((i for i in wrapped_coins if hasattr(i, "get_rate")), False):
-        contract = next(i for i in wrapped_coins if hasattr(i, "get_rate"))
-        rate_sig = right_pad(contract.get_rate.signature)
-    has_initial_A = hasattr(swap, "initial_A")
-    is_v1 = pool_data['lp_contract'] == "CurveTokenV1"
-
-    if hasattr(swap, "underlying_coins"):
-        registry.add_pool(
-            swap,
-            n_coins,
-            pool_token,
-            ZERO_ADDRESS,
-            rate_sig,
-            pack_values(wrapped_decimals),
-            pack_values(underlying_decimals),
-            has_initial_A,
-            is_v1,
-            {'from': alice}
-        )
-    else:
-        use_rates = [i['wrapped'] for i in pool_data['coins']] + [False] * (8 - n_coins)
-        registry.add_pool_without_underlying(
-            swap,
-            n_coins,
-            pool_token,
-            ZERO_ADDRESS,
-            rate_sig,
-            pack_values(underlying_decimals),
-            pack_values(use_rates),
-            has_initial_A,
-            is_v1,
-            {'from': alice}
-        )
-
-    yield registry
-
-
-@pytest.fixture(scope="module")
-def gauge_controller(GaugeControllerMock, alice):
-    yield GaugeControllerMock.deploy({'from': alice})
-
-
-@pytest.fixture(scope="module")
-def gauge_compound(LiquidityGaugeMock, accounts, gauge_controller, pool_token):
-    gauge = LiquidityGaugeMock.deploy(pool_token, {'from': accounts[0]})
-    gauge_controller._set_gauge_type(gauge, 1, {'from': accounts[0]})
-    yield gauge
-
-
-# helper functions
-
-@pytest.fixture(scope="session")
-def approx():
-
-    def _approx(a, b, precision=1e-10):
-        return 2 * abs(a - b) / (a + b) <= precision
-
-    yield _approx
